[수정] 댓글 구현으로 인한 게시글 불러오기 API 수정

+++ src/main/java/com/juno/normalapi/domain/vo/BoardVo.java
@@ -20,6 +20,6 @@ public class BoardVo {
     private String title;
     private String content;
     private String writer;
-    private int replyCount;
+    private Long replyCount;
     private LocalDateTime createdAt;
 }
